requirements.txt
asgiref==3.8.1
Django==5.0.3
django-filter==24.1
django-shortuuidfield==0.1.3
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.10.0
uritemplate==4.1.1
